REFACTOR odklad/2-osobne-udaje
